hospital-website/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==0.28.1
pinecone-client==2.2.4
langchain==0.0.340
pydantic==2.5.2 